+++ src/app/products/services/products.service.ts
@@ -1,25 +1,31 @@
 import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+
+import { Observable, throwError } from 'rxjs';
+import { catchError } from 'rxjs/operators';
+
 import { ProductModel } from '../models/product.model';
 
-import { productsData } from '../../shared/mocks';
-
-const productsList = productsData.map(item => new ProductModel(item.id, item.name, item.description, item.details, item.price, item.category, item.images, item.isAvailable));
-
-const productsListPromise = Promise.resolve(productsList);
 
 @Injectable()
 export class ProductsService {
+  private productsUrl = 'http://localhost:3000/products';
 
-  constructor() { }
+  constructor(
+    private http: HttpClient
+  ) { }
 
-  getProducts(): Promise<ProductModel[]> {
-    return productsListPromise;
+  getProducts(): Observable<ProductModel[]> {
+    return this.http
+      .get<ProductModel[]>(this.productsUrl)
+      .pipe(catchError((err) => throwError('Error in getProducts method')));
   }
 
   getProduct(id: number | string): Promise<ProductModel> {
     return this.getProducts()
+      .toPromise()
       .then(products => products.find(product => product.id === +id))
-      .catch(() => Promise.reject('Error in getProduct method'));
+      .catch((err) => Promise.reject(err));
   }
 
 }